Add local pipeline

+++ bagit_create/main.py
@@ -2,7 +2,7 @@ from .pipelines import invenio_v1
 from .pipelines import invenio_v3
 from .pipelines import opendata
 from .pipelines import indico
-
+from .pipelines import local
 
 import logging
 from fs import open_fs
@@ -16,6 +16,7 @@ def process(
     source,
     loglevel,
     target,
+    localsource,
     dry_run=False,
     alternate_uri=False,
     bibdoc=False,
@@ -50,6 +51,8 @@ def process(
             pipeline = invenio_v3.InvenioV3Pipeline(source)
         elif source == "indico":
             pipeline = indico.IndicoV1Pipeline("https://indico.cern.ch/")
+        elif source == "local":
+            pipeline = local.LocalV1Pipeline(localsource)
 
         # Save job details (Audit step 0)
         audit = [
@@ -59,39 +62,49 @@ def process(
             }
         ]
 
-        # Prepare folders
-        base_path, temp_files_path, name = pipeline.prepare_folders(source, recid)
+        if source == "local":
+            checksum = pipeline.get_folder_checksum(localsource)
+            base_path, name = pipeline.prepare_folders_ls(localsource, checksum)
+
+        else:
+            # Prepare folders
+            base_path, temp_files_path, name = pipeline.prepare_folders(source, recid)
 
         # Create bagit.txt
         pipeline.add_bagit_txt(f"{base_path}/bagit.txt")
 
-        # Get metadata
-        metadata, metadata_url, status_code, metadata_filename = pipeline.get_metadata(
-            recid
-        )
+        if source == "local":
+            pipeline.folder_creation(localsource, checksum, base_path)
 
-        # Save metadata file in the meta folder
-        pipeline.write_file(metadata, f"{base_path}/data/meta/{metadata_filename}")
-
-        # Parse metadata for files
-        files = pipeline.parse_metadata(f"{base_path}/data/meta/{metadata_filename}")
-
-        if dry_run is True:
-            # Create fetch.txt
-            pipeline.create_fetch_txt(files, f"{base_path}/fetch.txt", alternate_uri)
+            pipeline.create_manifests_ls(base_path)
+        
         else:
-            # Download files
-            pipeline.download_files(files, f"{base_path}/data/content")
+            # Get metadata
+            metadata, metadata_url, status_code, metadata_filename = pipeline.get_metadata(
+                recid
+            )
 
-        # Create sip.json
-        files = pipeline.create_bic_meta(
-            files, audit, metadata_filename, metadata_url, base_path
-        )
+            # Save metadata file in the meta folder
+            pipeline.write_file(metadata, f"{base_path}/data/meta/{metadata_filename}")
+
+            # Parse metadata for files
+            files = pipeline.parse_metadata(f"{base_path}/data/meta/{metadata_filename}")
+
+            if dry_run is True:
+                # Create fetch.txt
+                pipeline.create_fetch_txt(files, f"{base_path}/fetch.txt", alternate_uri)
+            else:
+                # Download files
+                pipeline.download_files(files, f"{base_path}/data/content")
+
+            # Create sip.json
+            files = pipeline.create_bic_meta(
+                files, audit, metadata_filename, metadata_url, base_path
+            )
+            pipeline.create_manifests(files, base_path)
         # an entry for "sip.json" gets added to files
 
         # Create manifest files
-        pipeline.create_manifests(files, base_path)
-
         pipeline.add_bag_info(base_path, f"{base_path}/bag-info.txt")
 
         # Verify created Bag
@@ -109,8 +122,9 @@ def process(
                 pipeline.delete_folder(base_path)
 
                 return {"status": 1, "errormsg": e}
-
-        pipeline.delete_folder(temp_files_path)
+        
+        if not source == "local":
+            pipeline.delete_folder(temp_files_path)
 
         logging.info("SIP successfully created")
 
@@ -126,7 +140,8 @@ def process(
     #  any created file and folder
     except Exception as e:
         logging.error(f"Job failed with error: {e}")
-        pipeline.delete_folder(temp_files_path)
+        if not source == "local": 
+            pipeline.delete_folder(temp_files_path)
         pipeline.delete_folder(base_path)
 
         return {"status": 1, "errormsg": e}
